move independent test to its own file
