+++ editor/src/components/inspector/common/inspector-end-to-end-tests.spec.browser2.tsx
@@ -1,10 +1,27 @@
 /* eslint-disable jest/expect-expect */
-import { fireEvent, RenderResult, screen } from '@testing-library/react'
+import { act, fireEvent, RenderResult, screen } from '@testing-library/react'
+import * as Prettier from 'prettier/standalone'
+import { PrettierConfig } from 'utopia-vscode-common'
+import { matchInlineSnapshotBrowser } from '../../../../test/karma-snapshots'
+import { FOR_TESTS_setNextGeneratedUids } from '../../../core/model/element-template-utils.test-utils'
+import { directory } from '../../../core/model/project-file-utils'
 import {
   BakedInStoryboardUID,
   BakedInStoryboardVariableName,
 } from '../../../core/model/scene-utils'
 import * as EP from '../../../core/shared/element-path'
+import {
+  ElementPath,
+  ProjectContents,
+  RevisionsState,
+  textFile,
+  textFileContents,
+  unparsed,
+} from '../../../core/shared/project-file-types'
+import { setFeatureEnabled } from '../../../utils/feature-switches'
+import { expectSingleUndoStep, selectComponentsForTest } from '../../../utils/utils.test-utils'
+import { contentsToTree } from '../../assets'
+import { SubduedBorderRadiusControlTestId } from '../../canvas/controls/select-mode/subdued-border-radius-control'
 import {
   EditorRenderResult,
   getPrintedUiJsCode,
@@ -16,33 +33,14 @@ import {
   TestScenePath,
   TestSceneUID,
 } from '../../canvas/ui-jsx.test-utils'
+import { createCodeFile } from '../../custom-code/code-file.test-utils'
+import { EditorAction } from '../../editor/action-types'
 import {
-  deleteSelected,
   selectComponents,
   sendLinterRequestMessage,
   updateFromCodeEditor,
 } from '../../editor/actions/action-creators'
-import { PrettierConfig } from 'utopia-vscode-common'
-import * as Prettier from 'prettier/standalone'
-import { act } from '@testing-library/react'
-import { contentsToTree } from '../../assets'
-import {
-  ElementPath,
-  ProjectContents,
-  RevisionsState,
-  textFile,
-  textFileContents,
-  unparsed,
-} from '../../../core/shared/project-file-types'
-import { directory } from '../../../core/model/project-file-utils'
 import { DefaultPackageJson, StoryboardFilePath } from '../../editor/store/editor-state'
-import { createCodeFile } from '../../custom-code/code-file.test-utils'
-import { matchInlineSnapshotBrowser } from '../../../../test/karma-snapshots'
-import { EditorAction } from '../../editor/action-types'
-import { expectSingleUndoStep, selectComponentsForTest } from '../../../utils/utils.test-utils'
-import { SubduedBorderRadiusControlTestId } from '../../canvas/controls/select-mode/subdued-border-radius-control'
-import { FOR_TESTS_setNextGeneratedUids } from '../../../core/model/element-template-utils.test-utils'
-import { setFeatureEnabled } from '../../../utils/feature-switches'
 import {
   ConditionalsControlSectionCloseTestId,
   ConditionalsControlSectionOpenTestId,
@@ -2365,53 +2363,6 @@ describe('inspector tests with real metadata', () => {
                   <div data-uid='ccc' data-testid='ccc'>bar</div>
                 ) /* this is a test */
                 // and another comment
-              }
-            </div>
-         `),
-      )
-    })
-    it('deleting a conditional branch replaces it with null', async () => {
-      FOR_TESTS_setNextGeneratedUids([
-        'skip1',
-        'skip2',
-        'skip3',
-        'skip4',
-        'skip5',
-        'skip6',
-        'conditional',
-      ])
-      const startSnippet = `
-        <div data-uid='aaa'>
-        {
-          [].length === 0 ? (
-            <div data-uid='bbb' data-testid='bbb'>foo</div>
-          ) : (
-            <div data-uid='ccc' data-testid='ccc'>bar</div>
-          )
-        }
-        </div>
-      `
-      const renderResult = await renderTestEditorWithCode(
-        makeTestProjectCodeWithSnippet(startSnippet),
-        'await-first-dom-report',
-      )
-
-      const targetPath = EP.appendNewElementPath(TestScenePath, ['aaa', 'conditional', 'bbb'])
-      await act(async () => {
-        await renderResult.dispatch([selectComponents([targetPath], false)], false)
-      })
-
-      await act(async () => {
-        await renderResult.dispatch([deleteSelected()], true)
-      })
-
-      expect(getPrintedUiJsCode(renderResult.getEditorState())).toEqual(
-        makeTestProjectCodeWithSnippet(`
-            <div data-uid='aaa'>
-              {
-                [].length === 0 ? null : (
-                  <div data-uid='ccc' data-testid='ccc'>bar</div>
-                )
               }
             </div>
          `),
